changes to remove warnings

+++ src/phys-boundary/calibrators/SensorDataCalibrator.cpp
@@ -17,11 +17,15 @@ void SensorDataCalibrator::applyCalibration(SensorData &data) {
 
 
 std::function<int(int)> IntFuncts::Quadratic(double a, double b, double c){
-    return [a,b,c](int x){return c + (b * x) + (a * x * x);};
+    return [a,b,c](int x){
+        return static_cast<int>(c + (b * x) + (a * x * x));
+    };
 }
 
 std::function<int(int)> IntFuncts::Linear(double m, double b){
-    return [m,b](int x){return b + x * m;};
+    return [m,b](int x){
+        return static_cast<int>(b + x * m);
+    };
 }
